+++ src/ReqRest/Builders/RequestUriBuilderExtensions.cs
@@ -11,17 +11,17 @@
     {
 
         /// <summary>
-        ///     Executes the specified <paramref name="configure"/> method which allows configuring the
-        ///     <see cref="IRequestUriBuilder.RequestUri"/> via an <see cref="UrlBuilder"/>.
+        ///     Modifies the current request URI by invoking the specified <paramref name="configure"/>
+        ///     function which can then modify the <see cref="IRequestUriBuilder.RequestUri"/>
+        ///     via an <see cref="UrlBuilder"/>.
         /// </summary>
         /// <typeparam name="T">The type of the builder.</typeparam>
         /// <param name="builder">The builder.</param>
         /// <param name="configure">
-        ///     A method which receives a new <see cref="UrlBuilder"/> instance created from the
+        ///     A function which receives a new <see cref="UrlBuilder"/> instance created from the
         ///     current <see cref="IRequestUriBuilder.RequestUri"/>.
-        /// 
-        ///     The <see cref="Uri"/> which was built with the provided <see cref="UrlBuilder"/> is
-        ///     used as the new <see cref="IRequestUriBuilder.RequestUri"/>.
+        ///     The URI which has been built by the returned <see cref="UrlBuilder"/> is used as the
+        ///     new <see cref="IRequestUriBuilder.RequestUri"/>.
         /// </param>
         /// <returns>The specified <paramref name="builder"/>.</returns>
         /// <exception cref="ArgumentNullException">
@@ -45,19 +45,17 @@
         }
 
         /// <summary>
-        ///     Executes the specified <paramref name="configure"/> function which allows configuring the
-        ///     <see cref="IRequestUriBuilder.RequestUri"/> via an <see cref="UrlBuilder"/>.
+        ///     Modifies the current request URI by invoking the specified <paramref name="configure"/>
+        ///     function which can then modify the <see cref="IRequestUriBuilder.RequestUri"/>
+        ///     via an <see cref="UrlBuilder"/>.
         /// </summary>
         /// <typeparam name="T">The type of the builder.</typeparam>
         /// <param name="builder">The builder.</param>
         /// <param name="configure">
         ///     A function which receives a new <see cref="UrlBuilder"/> instance created from the
         ///     current <see cref="IRequestUriBuilder.RequestUri"/>.
-        /// 
-        ///     The resulting <see cref="Uri"/> is used as the new <see cref="IRequestUriBuilder.RequestUri"/>.
-        ///
-        ///     Hint: You can simply return the <see cref="UrlBuilder"/>, because it gets implicitly converted
-        ///     to an <see cref="Uri"/>.
+        ///     The URI which has been built by the returned <see cref="UrlBuilder"/> is used as the
+        ///     new <see cref="IRequestUriBuilder.RequestUri"/>.
         /// </param>
         /// <returns>The specified <paramref name="builder"/>.</returns>
         /// <exception cref="ArgumentNullException">
@@ -68,6 +66,41 @@
         ///     The <see cref="Uri"/> which was built by the <paramref name="configure"/> function
         ///     had an invalid format.
         /// </exception>
+        [DebuggerStepThrough]
+        public static T ConfigureRequestUri<T>(this T builder, Func<UrlBuilder, UriBuilder?> configure) where T : IRequestUriBuilder
+        {
+            _ = builder ?? throw new ArgumentNullException(nameof(builder));
+            _ = configure ?? throw new ArgumentNullException(nameof(configure));
+
+            var urlBuilder = builder.RequestUri is null
+                ? new UrlBuilder()
+                : new UrlBuilder(builder.RequestUri);
+            var result = configure(urlBuilder);
+            return builder.SetRequestUri(result?.Uri);
+        }
+
+        /// <summary>
+        ///     Modifies the current request URI by invoking the specified <paramref name="configure"/>
+        ///     function which can then modify the <see cref="IRequestUriBuilder.RequestUri"/>
+        ///     via an <see cref="UrlBuilder"/>.
+        /// </summary>
+        /// <typeparam name="T">The type of the builder.</typeparam>
+        /// <param name="builder">The builder.</param>
+        /// <param name="configure">
+        ///     A function which receives a new <see cref="UrlBuilder"/> instance created from the
+        ///     current <see cref="IRequestUriBuilder.RequestUri"/>.
+        ///     The returned <see cref="Uri"/> is used as the new <see cref="IRequestUriBuilder.RequestUri"/>.
+        /// </param>
+        /// <returns>The specified <paramref name="builder"/>.</returns>
+        /// <exception cref="ArgumentNullException">
+        ///     * <paramref name="builder"/>
+        ///     * <paramref name="configure"/>
+        /// </exception>
+        /// <exception cref="UriFormatException">
+        ///     The <see cref="Uri"/> which was built by the <paramref name="configure"/> function
+        ///     had an invalid format.
+        /// </exception>
+        [DebuggerStepThrough]
         public static T ConfigureRequestUri<T>(this T builder, Func<UrlBuilder, Uri?> configure) where T : IRequestUriBuilder
         {
             _ = builder ?? throw new ArgumentNullException(nameof(builder));
@@ -76,18 +109,17 @@
             var urlBuilder = builder.RequestUri is null
                 ? new UrlBuilder()
                 : new UrlBuilder(builder.RequestUri);
-
             var uri = configure(urlBuilder);
             return builder.SetRequestUri(uri);
         }
 
         /// <summary>
-        ///     Sets the request <see cref="Uri"/> which is being built.
+        ///     Sets the request URI which is being built.
         /// </summary>
         /// <typeparam name="T">The type of the builder.</typeparam>
         /// <param name="builder">The builder.</param>
         /// <param name="requestUri">
-        ///     A string which represents the request <see cref="Uri"/> to be used.
+        ///     A string which represents the request URI to be used.
         ///     This can be <see langword="null"/>.
         /// </param>
         /// <param name="uriKind">
@@ -103,12 +135,12 @@
                 builder.SetRequestUri(requestUri is null ? null : new Uri(requestUri, uriKind));
 
         /// <summary>
-        ///     Sets the request <see cref="Uri"/> which is being built.
+        ///     Sets the request URI which is being built.
         /// </summary>
         /// <typeparam name="T">The type of the builder.</typeparam>
         /// <param name="builder">The builder.</param>
         /// <param name="requestUri">
-        ///     The request <see cref="Uri"/> used.
+        ///     The request URI used.
         ///     This can be <see langword="null"/>.
         /// </param>
         /// <returns>The specified <paramref name="builder"/>.</returns>